ChromaEmitters/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ChromaEmitters"
version = "0.3.9"
description = "Device output layer for Chromaspace frames"
readme = "docs/README.md"
requires-python = ">=3.8"
dependencies = ["tUilKit", "PyYAML"]

[tool.setuptools.packages.find]
where = ["src"]




